+++ src/core/keyboard/constructor.js
@@ -15,22 +15,19 @@ import {
 } from './regex';
 import puncCodeMap from './punc-code-map.json';
 
-// On macs the metaKey is triggered. We want them to be treated equivalently?
-// So the value of the ctrlKey property should be (ctrlKey || metaKey
 const modifierPropsMap = {
   [CTRL]: 'ctrlKey',
   [ALT]: 'altKey',
   [SHIFT]: 'shiftKey',
-  [META]: 'ctrlKey', // Point meta key to ctrl key
+  [META]: 'metaKey',
 };
-
-// const modifiersArray = Object.keys(modifierPropsMap);
 
 const protoKbdCommand = {
   key: null,
   ctrlKey: false,
   altKey: false,
   shiftKey: false,
+  metaKey: false,
 };
 
 // Returns a bool indicating whether this input is valid
@@ -56,8 +53,7 @@ export function kbdCommand(input) {
     return kbdCommandEvent(input);
   }
   throw new TypeError(
-    "Command input isn't the right type. KbdCommand requires a valid command\
-    string or object",
+    "Command input isn't the right type. KbdCommand requires a valid command string or object",
   );
 }
 
@@ -104,7 +100,7 @@ function kbdCommandEvent(event) {
   return makeKbdCommandFromEvent(event);
 }
 
-const isSingleKey = c => (!c.ctrlKey && !c.shiftKey && !c.altKey);
+const isSingleKey = c => (!c.ctrlKey && !c.shiftKey && !c.altKey && !c.metaKey);
 const isFinalKeyModifier = c => c.key === CONTROL
   || c.key === CTRL
   || c.key === META
@@ -120,7 +116,8 @@ function makeKbdCommandFromEvent({
 }) {
   const command = {
     key: capitalizeFirstLetter(key),
-    ctrlKey: ctrlKey || metaKey,
+    ctrlKey,
+    metaKey,
     altKey,
     shiftKey,
   };