+++ src/core/pages/popup/utils/dom.js
@@ -15,7 +15,6 @@ import {
   TAB_MUTED_CLASSNAME,
   TAB_ACTIVE_ID,
   OTHER_WINDOW_TAB_TYPE,
-  SESSION_TYPE,
   TAB_TYPE,
   DEL_CIRCLE_SVG_PATH,
   TAB_DELETE_BTN_CLASSNAME,
@@ -55,8 +54,8 @@ function clearChildren(node) {
 // Given a string, return true if one string is too long and should apply
 function shouldWordBreak(str) {
   const MAX_LENGTH = 40;
-  // eslint-disable-next-line arrow-body-style
-  const longestWordLength = (acc, s) => s.length > acc
+  // eslint-disable-next-line no-confusing-arrow
+  const longestWordLength = (acc, s) => (s.length > acc)
     ? s.length
     : acc;
   const longest = str.split(/\s/).reduce(longestWordLength, 0);
@@ -71,14 +70,6 @@ function inlineOpts(setting) {
     return Object.assign({ [c]: setting[c] }, acc);
   }, {});
 }
-
-const colorTypeMap = {
-  [TAB_TYPE]: 'tabColor',
-  [OTHER_WINDOW_TAB_TYPE]: 'otherWindowTabColor',
-  [SESSION_TYPE]: 'recentlyClosedTabColor',
-  [BOOKMARK_TYPE]: 'bookmarkColor',
-  [HISTORY_TYPE]: 'historyColor',
-};
 
 export function tabToTag(getState) {
   const { color: colorSettings } = getState();
@@ -111,10 +102,6 @@ export function tabToTag(getState) {
     // Since favicon url of bookmarks isn't readily available,
     // check the type and assign all bookmarks to the static svg
     // for now.
-    const tabInline = {};
-    if (colorTypeMap[type] in inline) {
-      tabInline.borderColor = inline[colorTypeMap[type]];
-    }
     let favIconLink;
     switch (type) {
       case HISTORY_TYPE:
@@ -128,7 +115,6 @@ export function tabToTag(getState) {
         break;
     }
     const ctoOpts = {
-      tabInline,
       tabUrlSize,
       tabTitleSize,
       showRecentlyClosed,
@@ -174,7 +160,6 @@ function createTabObject({
 }, opts) {
   const {
     wordBreak,
-    tabInline,
     tabUrlSize,
     tabTitleSize,
     showVisualDeleteTabButton,
@@ -187,9 +172,6 @@ function createTabObject({
   tabObjectNode.setAttribute('tabindex', '0');
   tabObjectNode.classList.add('tab-object');
   tabObjectNode.classList.add(type);
-  if (tabInline && tabInline.borderColor) {
-    tabObjectNode.style.borderColor = tabInline.borderColor;
-  }
   // Declare id used for switching
   tabObjectNode.setAttribute('data-id', dataId);
   // Declare data type to know what to paint
@@ -207,7 +189,7 @@ function createTabObject({
 
   const favIconNode = d.createElement('img');
   favIconNode.setAttribute('src', favIconLink);
-  favIconNode.onerror = handleBadSvg;
+  favIconNode.addEventListener('error', handleBadSvg);
 
   iconContainerNode.appendChild(favIconNode);
 
@@ -273,9 +255,9 @@ function createTabObject({
   return tabObjectNode;
 }
 
-function handleBadSvg() {
-  badFavIconCache().add(this.src);
-  this.src = favIconFallback;
+function handleBadSvg(event) {
+  badFavIconCache().add(event.currentTarget.src);
+  event.currentTarget.src = favIconFallback;
 }
 
 // Get focused node's position relative to the current scrolled view
@@ -367,6 +349,8 @@ export function removeElementFromTabList(element, wasClicked) {
   }
 }
 
+// Changes both the type classname (for styling)
+// and the type prop in the dataset
 export function repaintElementWithType(element, typeToAdd) {
   const { type: typeToRemove } = element.dataset;
   element.classList.remove(typeToRemove);
@@ -398,4 +382,34 @@ export function selectNodeText(nodeToCopy) {
   const range = d.createRange();
   range.selectNode(nodeToCopy);
   window.getSelection().addRange(range);
+}
+
+// This is dependent on the popup css file being called "index.css" exactly
+const popupCssStyleSheet = [...d.styleSheets].find(s => /index.css$/.test(s.href));
+
+export function setStyleSheetRule(selectorName, key, value) {
+  const rules = popupCssStyleSheet.cssRules;
+  const rule = [...rules].find(r => r.selectorText === selectorName);
+  if (!rule) {
+    console.error(`Can't find CSS Selector ${selectorName}`);
+    return;
+  }
+  rule.style.setProperty(key, value);
+}
+
+// Finds the first element in tabList that has type
+// this should only be called if recently closed tabs are always shown at the
+// bottom of the tab list. If no recently recently closed tabs found,
+// returns undefined
+export function findLastElType(type, domNodes) {
+  let last;
+  // Iterate the array in reverse order
+  for (let i = domNodes.length - 1; i >= 0; i -= 1) {
+    const el = domNodes[i];
+    if (el.dataset.type !== type) {
+      return last;
+    }
+    last = el;
+  }
+  return last;
 }